refactor: 이름에 대한 유효 검사 책임을 다른 객체로 분리
